+++ common/src/main/java/io/homo/superresolution/core/graphics/impl/GpuObject.java
@@ -1,5 +1,5 @@
 package io.homo.superresolution.core.graphics.impl;
 
 public interface GpuObject {
-    int handle();
+    long handle();
 }
